+++ src/games/super-mario-land/src/interfaces/box.interface.ts
@@ -1,0 +1,8 @@
+export interface IBoxConstructor {
+  scene: Phaser.Scene;
+  content: any;
+  x: number;
+  y: number;
+  texture: string;
+  frame?: string | number;
+}