+++ src/games/super-mario-land/src/interfaces/brick.interface.ts
@@ -1,0 +1,8 @@
+export interface IBrickConstructor {
+  scene: Phaser.Scene;
+  value: number;
+  x: number;
+  y: number;
+  texture: string;
+  frame?: string | number;
+}