+++ src/games/super-mario-land/src/interfaces/collectible.interface.ts
@@ -1,0 +1,8 @@
+export interface ICollectibleConstructor {
+  scene: Phaser.Scene;
+  points: number;
+  x: number;
+  y: number;
+  texture: string;
+  frame?: string | number;
+}