+++ src/games/super-mario-land/src/interfaces/platform.interface.ts
@@ -1,0 +1,8 @@
+export interface IPlatformConstructor {
+  scene: Phaser.Scene;
+  tweenProps: any;
+  x: number;
+  y: number;
+  texture: string;
+  frame?: string | number;
+}