+++ src/games/super-mario-land/src/interfaces/portal-destination.interface.ts
@@ -1,0 +1,5 @@
+export interface IPortalDestination {
+  x: number;
+  y: number;
+  dir: string;
+}