+++ src/games/super-mario-land/src/interfaces/portal.interface.ts
@@ -1,0 +1,10 @@
+import { IPortalDestination } from './portal-destination.interface';
+
+export interface IPortalConstructor {
+  scene: Phaser.Scene;
+  spawn: IPortalDestination;
+  x: number;
+  y: number;
+  width?: number;
+  height?: number;
+}