+++ CHANGELOG.md
@@ -33,6 +33,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Refactor `flappy-bird`
 - Refactor `snake`
 - Refactor `space-invaders`
+- Refactor `super-mario-land`
 
 ---